refactor: address review, add dark-mode colourblind variants
- Extract ThemeContext/useTheme into src/contexts/theme.js so components no
  longer import from the app entrypoint (index.js has module-level render side
  effects); fixes the circular-dependency review comments.
- Give the colourblind palette lightened dark-mode variants so classes keep
  their visual punch on the dark background (all >= 3:1 contrast).
- Add aria-label + type="button" to the colourblind toggle for screen readers.
- Document why DecisionTree uses the saturated shade (colours are rendered as
  text there, not fills).
